fix: Calculate zoom ratio before rendering svg (#3616)
SvgComponent currently renders in lower-than-native resolution when
`game.camera.viewfinder.zoom` is larger than 1. This is addressed by
taking a ratio of `canvas.getDestinationClipBounds().size` and
`canvas.getLocalClipBounds().size`, and applying it to the existing
canvas scaling.

I'm still not quite sure how well this would work with `pixelRatio`, but
I've verified that when running the "render sharply" test which has the
device pixel ratio set to 3, the canvas size ratio mentioned above is 1,
and not rendering in excessive resolution.

The test uses a `Transform.scale` to simulate viewfinder zoom. I've also
verified that when running without the patch, this test generates a
blurry image:
|before|after|
|-|-|

|![Screenshot_20250604_073801](https://github.com/user-attachments/assets/1d4f7db0-6bcf-456c-a740-6b3f9a4b0f72)|![image](https://github.com/user-attachments/assets/476e774f-8547-48a5-9cf6-31f9476ec55b)|

---------

Co-authored-by: Lukas Klingsbo <lukas.klingsbo@gmail.com>
Co-authored-by: Lukas Klingsbo <me@lukas.fyi>