feat: Expose additional flame_jenny Command classes (#3641)
It is possible to take the name of a `Node` in a `YarnProject` and
inspect the `DialogueEntry` objects that its iterator returns. However,
even though `DialogueEntry` objects can be either a `DialogueLine`, a
`DialogueChoice`, or a `Command`, there is no way to actually reference
a `Command`! `UserDefinedCommand` exists as a concrete implementation of
`Command`, but there are other `Command` types that someone might be
interested in. For my purposes, the `JumpCommand` and `VisitCommand`
both are `Command` types that allow for the user to leave to other
nodes. It would be convenient to be able to recursively follow these
commands to their new nodes and also perform the same introspection that
I would like to do the original node. This PR just exposes these
additional, pre-existing classes for external use.