fix: Attach layout listeners to new children (#3648)
Prior to this fix, when adding a child to the `LayoutComponent`,
resizing the child wouldn't cause the `LayoutComponent` to resize. This
was because the children size listeners were only attached on
instantiation of `LayoutComponent` *and* during setting of `shrinkWrap`.

This fix attaches the size listeners to a single child when it is added
to `LayoutComponent` via `onChildrenChanged`.
`setupSizeListeners` is not used because it iterates through all the
children, and is unnecessary.