fix: Update flame_3d to support both old and newer Flutter APIs (#3663)
Several APIs started to return non-nullable types instead of null to
indicate errors. However we want to support both versions. If we do not
add `!`, it will fail on some versions, and if we do add `!`, the lint
will complain on others:

<img width="2428" height="1162" alt="image"
src="https://github.com/user-attachments/assets/73ee6679-4017-4c2b-a5d3-510d63ce5f27"
/>

This is a compromise, using a hacky "unwrap" method (a line-level
suppression would be fine; the unwrap just centralizes the hack in one
function for both methods).