fix: Fix color to byte conversion for shaders (#3700)
Fix color to byte conversion for shaders.

Why hadn't we noticed this before?
Everything in the example uses the `ColorTexture` to create an in-memory
texture with a solid color. That uses a different method to create
texture data using one byte per pixel:

```
        Uint32List.fromList(
          List.filled(
            width * height,
            // Convert to a 32 bit value representing this color.
            (color.a * 255).round() << 24 |
                (color.r * 255).round() << 16 |
                (color.g * 255).round() << 8 |
                (color.b * 255).round(),
          ),
        ).buffer.asByteData(),
```
As for the lights, they already used `setColor`, which works. But when
setting an albedo color instead of albedo texture, it would have its own
method that included an ad-hoc cache. That method was not wiring the
bytes in the correct order, causing the issue. I changed it to use
`setColor`, which is much nicer, and replaced the cache with a
global-level cache for mapping the immutable Colors onto bytes. That way
we get the benefit of the cache for lights too.

Before:
<img width="1686" height="1498" alt="image"
src="https://github.com/user-attachments/assets/58cb9872-ba8e-49dc-9fa0-5fb685925ebc"
/>

After:
<img width="1666" height="1470" alt="image"
src="https://github.com/user-attachments/assets/70ba73d5-2e74-4780-ad21-1d6869441b5c"
/>

Co-authored-by: Lukas Klingsbo <me@lukas.fyi>