docs: Remove the "Flame web" section in platforms.md (#3705)
Newer versions of Flutter (3.27+ according to the last link below)
default to the CanvasKit renderer, and following the instructions in
that documentation section to pass the `--web-renderer` results in the
following error (on Flame 1.32 & Flutter 3.35.2 which are the latest
versions as of this writing):
```
$ flutter build web --release --web-renderer canvaskit
Could not find an option named "--web-renderer".

Run 'flutter -h' (or 'flutter <command> -h') for available flutter commands and options.
```

Since this section only talked about this quirk I think it's better to
remove it altogether. If there's a reason to keep it and for users of
older versions of Flutter, a note on this can be added instead.

Addresses https://github.com/flame-engine/flame/issues/3706

See Also:

- https://github.com/flutter/flutter/issues/145954

- The bottom of
https://docs.google.com/document/d/1DGamHsa2lz_Qtgfrfa3j3fRaEopJXc7tCFVM1TQlck8/edit?resourcekey=0-IjkqrCoo_EcRRS2bPYKI8w&tab=t.0

- https://stackoverflow.com/a/79522746