perf: `addAll` shouldn't create unnecessary growing lists (#3737)
This PR removes the sometimes unnecessary creation of a growing list
from `addAll`.

> "I'm adding hundreds to thousands of components onto the screen in a
short period of time. Switching to .add from .addAll cut down on these
CPU-consuming functions:
> 
> _GrowableList.add, _allocateData, _growToNextCapacity, length=
> 
> Even if you're not adding components that need async onLoads, the
engine is always instantiating a future list and > adding to it."

Later on we should experiment with list comprehensions and see if they
create a statically sized list.