docs: Add console with backtick on flame_console example (#3743)
Add console with backtick on flame_console example

This makes the example a bit more ergonomic to test changes to the
console by providing a key bind, backtick, for the console. This is also
much more in line with how actual games would use this feature instead
of a FAB (for example, on Bethesda games and many others, backtick
brings up the console), and in particular will teach how to properly
propagate events so both in-game keyboard handling _and_ the console
keyboard handlers can co-exist.