refactor: Re-organize internal event imports (#3742)
Re-organize internal event files using the `events.dart` export.

This changes the imports on all files to use the exposed `events.dart`
export - this will vastly help us to catch issues of forgetting to
expose things. If anything is imported from `src/`, we know it is not
exposed by Flame.

This should be absolutely no-op whatsoever - just organizational.