feat: TextBoxComponent updateBounds on set boxConfig (#3777)
Currently, when `boxConfig.maxWidth` is changed, TextBoxComponent does
not reflow its text. This PR fixes this.

Note that this PR includes a non-standard control in its example. This
is intentional because the point of this PR is to support the change to
an existing instantiated component. Using a standard Dashbook control
obviates the point of this PR. Once
https://github.com/bluefireteam/dashbook/issues/128 is done, this will
be migrated.

---------

Co-authored-by: Lukas Klingsbo <me@lukas.fyi>