feat: Linear layout component text box component (#3779)
Modifies LinearLayoutComponent such that when oriented vertically and
set to CrossAxisAlignment.stretch, whenever it encounters a child that
is a TextBoxComponent, sets its `boxConfig.maxWidth` to the appropriate
width.

Currently dependent on https://github.com/flame-engine/flame/pull/3777,
so that has to be merged first before this.


This is only a minor breaking change. If you were using a
TextBoxComponent inside LinearLayoutComponent, chances are you had to do
weird things to make it work as you expect. You'd simply have to remove
the workaround code you wrote for it.

---------

Co-authored-by: Lukas Klingsbo <me@lukas.fyi>